# api/openapi.yaml
openapi: 3.0.3
info:
  title: NinShow APP
  description: |-
    This is the NinShow APP API documentation.
  termsOfService: https://localhost:5555
  contact:
    email: kotaro.otaka@example.com
  license:
    name: MIT
  version: 0.0.1
...
servers:
  - url: http://localhost:5555
...
paths:
...
  /idp/signup:
    post:
      tags:
        - IdP
      summary: Sign Up
      description: Sign Up
      operationId: idpSignup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdPSignupRequestSchema'
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
  /idp/signin:
    post:
      tags:
        - IdP
      summary: Sign In
      description: Sign In
      operationId: idpSignin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdPSigninRequestSchema'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /idp/oidc/login:
    get:
      tags:
        - IdP
      summary: OpenID Connect Login
      description: OpenID Connect Login
      operationId: idpOIDCLogin
      parameters:
        - name: op
          in: query
          required: true
          description: op
          schema:
            type: string
            enum:
              - zitadel
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /idp/oidc/callback:
    get:
      tags:
        - IdP
      summary: OpenID Connect Callback
      description: OpenID Connect Callback
      operationId: idpOIDCCallback
      parameters:
        - name: code
          in: query
          required: true
          description: code
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        - name: state
          in: query
          required: true
          description: state
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
      responses:
        '200':
          description: OK
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
  /op/.well-known/openid-configuration:
    get:
      tags:
        - OP
      summary: OpenID Provider Configuration
      description: OpenID Provider Configuration
      operationId: opOpenIDConfiguration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPOpenIDConfigurationResponseSchema'
        '500':
          description: Internal Server Error
  /op/authorize:
    get:
      tags:
        - OP
      summary: Authentication Request
      description: |
        Authentication Request
        ref: https://openid-foundation-japan.github.io/rfc6749.ja.html#code-authz-req
      operationId: opAuthorize
      parameters:
        - name: response_type
          in: query
          required: true
          description: response_type
          schema:
            type: string
            enum:
              - code
        - name: scope
          in: query
          required: true
          description: scope
          schema:
            type: string
            example: openid profile email
        - name: client_id
          in: query
          required: true
          description: client_id
          schema:
            type: string
            format: uri
            example: http://localhost:8080
        - name: redirect_uri
          in: query
          required: true
          description: http://localhost:8080/rp/callback
          schema:
            type: string
            format: uri
            example: http://localhost:8080/rp/callback
        - name: state
          in: query
          required: false
          description: state
          schema:
            type: string
            example: state
        - name: nonce
          in: query
          required: false
          description: nonce
          schema:
            type: string
            example: nonce
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080/rp/callback
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: error
                    enum:
                      - invalid_request
                      - invalid_scope
                    example: invalid_request
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: error
                    enum:
                      - unauthorized_client
                    example: unauthorized_client
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: error
                    enum:
                      - access_denied
                    example: access_denied
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: error
                    enum:
                      - server_error
                    example: server_error
  /op/login:
    get:
      tags:
        - OP
      summary: OP Login
      description: OP Login
      operationId: opLoginView
      parameters:
        - name: auth_request_id
          in: query
          required: true
          description: auth request id
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              schema:
                type: string
                example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
    post:
      tags:
        - OP
      summary: OP Login
      description: OP Login
      operationId: opLogin
      requestBody:
        description: OP Token request body 
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OPLoginRequestSchema'
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080/op/callback
        '500':
          description: Internal Server Error
  /op/callback:
    get:
      tags:
        - OP
      summary: OP Callback
      description: OP Callback
      operationId: opCallback
      parameters:
        - name: id
          in: query
          required: true
          description: id
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080/rp/callback
        '500':
          description: Internal Server Error
  /op/token:
    post:
      tags:
        - OP
      summary: OP Token Request
      description: |
        OP Token Request
        ref: https://openid-foundation-japan.github.io/rfc6749.ja.html#token-req
        ref: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#TokenRequest
      operationId: opToken
      requestBody:
        description: OP Token request body 
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OPTokenRequestSchema'
      responses:
        '200':
          description: OK
          headers:
            Cache-Control: 
              schema:
                type: string
                example: no-store
            Pragma:
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPTokenResponseSchema'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: error
                    example: invalid_request
        '500':
          description: Internal Server Error
  /op/userinfo:
    get:
      tags:
        - OP
      summary: UserInfo Request
      description: https://openid.net/specs/openid-connect-core-1_0.html#UserInfo
      operationId: opUserinfo
      security:
        - Bearer: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPUserInfoResponseSchema'
        '500':
          description: Internal Server Error
  /op/revoke:
    post:
      tags:
        - OP
      summary: OP Revocation Request
      description: https://openid.net/specs/openid-connect-core-1_0.html#Revocation
      operationId: opRevoke
      security:
        - Bearer: []
      requestBody:
        description: OP Revoke request body 
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OPRevokeRequestSchema'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: error
                    example: invalid_request
        '500':
          description: Internal Server Error
  /op/certs:
    get:
      tags:
        - OP
      summary: OP JWK Set
      description: https://openid-foundation-japan.github.io/rfc7517.ja.html
      operationId: opCerts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OPJWKSetResponseSchema'
        '500':
          description: Internal Server Error
  /rp/login:
    get:
      tags:
        - RP
      summary: RP Login
      description: RP Login
      operationId: rpLogin
      responses:
        '302':
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080/op/authorize
            Set-Cookie:
              schema: 
                type: string
                example: state=xxxxxxxxx; Path=/; HttpOnly
        '500':
          description: Internal Server Error
  /rp/callback:
    get:
      tags:
        - RP
      summary: RP Callback
      description: RP Callback
      operationId: rpCallback
      parameters:
        - name: code
          in: query
          required: true
          description: code
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        - name: state
          in: query
          required: true
          description: state
          schema:
            type: string
            example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
      responses:
        '200':
          description: OK
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    IdPSignupRequestSchema:
      type: object
      properties:
        name:
          type: string
          description: name
          example: name
        email:
          type: string
          format: email
          description: email
          example: example@example.com
        password:
          type: string
          description: password
          example: password
      required:
        - name
        - email
        - password
    IdPSigninRequestSchema:
      type: object
      properties:
        email:
          type: string
          format: email
          description: email
          example: example@example.com
        password:
          type: string
          description: password
          example: password
      required:
        - email
        - password
    OPOpenIDConfigurationResponseSchema:
      type: object
      properties:
        issuer:
          type: string
          description: http://localhost:8080/op
          format: uri
          example: http://localhost:8080/op
        authorization_endpoint:
          type: string
          description: http://localhost:8080/op/authorize
          format: uri
          example: http://localhost:8080/op/authorize
        token_endpoint:
          type: string
          description: http://localhost:8080/op/token
          format: uri
          example: http://localhost:8080/op/token
        userinfo_endpoint:
          type: string
          description: http://localhost:8080/op/userinfo
          format: uri
          example: http://localhost:8080/op/userinfo
        jwks_uri:
          type: string
          description: http://localhost:8080/op/certs
          format: uri
          example: http://localhost:8080/op/certs
        revocation_endpoint:
          type: string
          description: http://localhost:8080/op/revoke
          format: uri
          example: http://localhost:8080/op/revoke
      required:
        - issuer
        - authorization_endpoint
        - token_endpoint
        - userinfo_endpoint
        - jwks_uri
        - revocation_endpoint
    OPLoginRequestSchema:
      type: object
      properties:
        id:
          type: string
          description: id
          example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        email:
          type: string
          format: email
          description: email
          example: example@example.com
        password:
          type: string
          description: password
          example: password
      required:
        - id
        - email
        - password
...
    OPUserInfoResponseSchema:
      type: object
      description: https://openid.net/specs/openid-connect-core-1_0.html#UserInfoResponse
      properties:
        sub:
          type: string
          description: sub
          example: sub
        profile:
          type: string
          description: profile
          example: profile
        email:
          type: string
          description: email
          example: email
      required:
        - sub
...